ecommerce_microservice/admin/requirements.txt
Django==3.2.4
django-cors-headers==3.7.0
djangorestframework==3.12.4
mysqlclient==2.0.3
pika==1.2.0
django-mysql==3.12.0
python-dotenv==0.15.0